+++ srv/http/addonsdl.sh
@@ -31,6 +31,12 @@ versionlist=$( echo "$addonslist" | grep 'version.*=>' | cut -d "'" -f 4 )
 versionredis=$( redis-cli hget addons addo )
 
 if [[ $versionlist != $versionredis ]]; then
+	if (( $( df | grep '/$' | awk '{print $4}' ) < 1000 )); then
+		# get directory size if enough after delete files
+		(( $( du /var/cache/pacman/pkg | awk '{print $1}' ) < 1000 )) && exit 2
+		rm /var/cache/pacman/pkg/*
+		(( $( df | grep '/$' | awk '{print $4}' ) < 1000 )) && exit 2
+	fi
 	curl -s -v -X POST 'http://localhost/pub?id=addons' -d 1
 	wget -qN $installurl -P /srv/http
 	chmod 755 /srv/http/install.sh || exit 1