+++ srv/http/assets/js/addons.js
@@ -33,11 +33,9 @@ $( '.btnin' ).each( function() {
 			, cancel   : 1
 			, ok       : function() {
 				branch = $( '#infoTextbox' ).val() +' -b';
-				option = $thisbtn.attr( 'option' );
+				option = addons[ alias ].option;
 				if ( option ) {
 					j = 0;
-					option = option.replace( /'/g, '"' ); // double quote only for JSON.parse()
-					option = JSON.parse( option );
 					getoptions();
 				} else {
 					formtemp();
@@ -54,27 +52,30 @@ $( '.boxed-group .btn' ).click( function () {
 	type = $thisbtn.text().trim();
 	title = $thisbtn.parent().prev().prev().find( 'span' ).text();
 	
-	if ( type === 'Link' ) {
+	if ( alias === 'bash' ) {
+		option = addons[ alias ].option;
+		j = 0;
+		getoptions();
+	} else if ( type === 'Link' ) {
 		window.open( $thisbtn.prev().find( 'a' ).attr( 'href' ), '_blank' );
-		return
-	}
-	info( {
-		  title  : title
-		, message: type +'?'
-		, cancel : 1
-		, ok     : function () {
-			option = $thisbtn.attr( 'option' );
-			if ( option ) {
-				j = 0;
-				option = option.replace( /'/g, '"' ); // double quote only for JSON.parse()
-				option = JSON.parse( option );
-				getoptions();
-			} else {
-				formtemp();
+	} else if ( type === 'Update' ) {
+		formtemp();
+	} else {
+		info( {
+			  title  : title
+			, message: type +'?'
+			, cancel : 1
+			, ok     : function () {
+				option = addons[ alias ].option;
+				if ( option ) {
+					j = 0;
+					getoptions();
+				} else {
+					formtemp();
+				}
 			}
-		}
-	} );
-	if ( alias === 'bash' ) $( '#infoOk' ).click();
+		} );
+	}
 } );
 $( '.thumbnail' ).click( function() {
 	$sourcecode = $( this ).prev().find('form a').attr( 'href');