+++ changelog.md
@@ -1,3 +1,9 @@
+### 20171111
+- Check available disk space before update
+- Show available disk space
+- Notify installation to other clients
+- Change options format to PHP array
+
 ### 20171025
 - Add update revision of each addon
 - Add long-press Install for branch testing